requirements.txt
PyQt5>=5.15
Pillow>=10.0
numpy>=1.24
matplotlib>=3.7
